Make changes to onboarding image
