+++ src/App.css
@@ -122,6 +122,12 @@ a:hover {
     text-align: right;
   }
 
+  .modal-body {
+    padding: 0;
+    margin: 0;
+    border: 0;
+  }
+
   .modal-desc {
     text-align: center;
     padding: 0 30px 0 30px;
@@ -153,7 +159,7 @@ a:hover {
   }
 
   .modal-img {
-    margin: 5px 0 5px 0;
+    margin: 30px 0 30px 0;
   }
 
   .user-form {
@@ -179,18 +185,6 @@ a:hover {
     background: white;
     border: 1px solid #B5C5CA;;
     border-radius: 4px;
-  }
-
-  .fade-in {
-    visibility: visible;
-    opacity: 1;
-    transition: visibility 0s linear 0s, opacity 300ms;
-  }
-
-  .fade-out {
-    visibility: hidden;
-    opacity: 0;
-    transition: visibility 0s linear 300ms, opacity 300ms; 
   }
 
   @keyframes in {
