+++ src/MainWindow/HomePage/index.jsx
@@ -6,7 +6,6 @@ import InitScanForm from "./InitScanForm";
 import "./HomePage.scss";
 import services from "../../services";
 import { urlErrorCodes, urlErrorTypes } from "../../common/constants";
-import OnboardingComponent from "./OnboardingComponent";
 
 const HomePage = ({ appVersion, setCompletedScanId }) => {
   const navigate = useNavigate();
@@ -17,15 +16,6 @@ const HomePage = ({ appVersion, setCompletedScanId }) => {
   const [email, setEmail] = useState(''); 
   const [name, setName] = useState('');
   const [autoSubmit, setAutoSubmit] = useState(false);
-  const [userInputErrorMessage, setUserInputErrorMessage] = useState('');
-  const [dataExistStatus, setDataExistStatus] = useState(null);
-
-  useEffect(() => {
-    window.services.userDataExists((status) => {
-      console.log(status);
-      setDataExistStatus(status)
-    })
-  }, []);
 
   useEffect(() => {
     const getUserData = async () => {
@@ -39,10 +29,8 @@ const HomePage = ({ appVersion, setCompletedScanId }) => {
       }
     }
 
-    if (dataExistStatus == "exists") {
-      getUserData();
-    }
-  }, [dataExistStatus]);
+    getUserData();
+  });
 
   const isValidHttpUrl = (input) => {
     const regexForUrl = new RegExp("^(http|https):/{2}.+$", "gmi");
@@ -95,65 +83,38 @@ const HomePage = ({ appVersion, setCompletedScanId }) => {
     navigate("/error");
   };
 
-  const handleSetUserData =  (event) =>  {
-    event.preventDefault();
-
-    if (!services.isValidEmail(email)) {
-      setUserInputErrorMessage('Please enter a valid email.'); 
-      return;
-    }
-
-    window.services.setUserData({name: name, email: email, autoSubmit: true});
-    setDataExistStatus("exists");
-  }
-
-  if (dataExistStatus === "doesNotExist") {
-    return (
-      <OnboardingComponent 
-        handleSetUserData={handleSetUserData} 
-        setName={setName}
-        setEmail={setEmail}
-        userInputErrorMessage={userInputErrorMessage}
-      />
-    )
-  }
-
-  if (dataExistStatus === "exists") {
-    return (
-      <div id="home-page">
-        <div id="home-page-main">
-          <img
-            id="a11y-logo"
-            src={a11yLogo}
-            alt="Logo of the GovTech Accessibility Enabling Team"
-          />
-          <h1 id="app-title">Accessibility Site Scanner</h1>
-          <InitScanForm
-            startScan={startScan}
-            prevUrlErrorMessage={prevUrlErrorMessage}
-          />
-          {autoSubmit && 
-            (<div id="user-details">
-              <div>{name}</div>
-              <div>{email}</div>
-            </div>)
-          }
-        </div>
-        <div id="home-page-footer">
-          <img
-            id="app-illustration"
-            src={appIllustration}
-            alt="Illustration showing people with sight, hearing, motor and cognitive disabilities"
-          />
-          <span id="footer-text">
-            Version {appVersion} | Built by GovTech Accessibility Enabling Team
-          </span>
-        </div>
+  return (
+    <div id="home-page">
+      <div id="home-page-main">
+        <img
+          id="a11y-logo"
+          src={a11yLogo}
+          alt="Logo of the GovTech Accessibility Enabling Team"
+        />
+        <h1 id="app-title">Accessibility Site Scanner</h1>
+        <InitScanForm
+          startScan={startScan}
+          prevUrlErrorMessage={prevUrlErrorMessage}
+        />
+        {autoSubmit && 
+          (<div id="user-details">
+            <div>{name}</div>
+            <div>{email}</div>
+          </div>)
+        }
       </div>
-    );
-  }
-
-  return null;
+      <div id="home-page-footer">
+        <img
+          id="app-illustration"
+          src={appIllustration}
+          alt="Illustration showing people with sight, hearing, motor and cognitive disabilities"
+        />
+        <span id="footer-text">
+          Version {appVersion} | Built by GovTech Accessibility Enabling Team
+        </span>
+      </div>
+    </div>
+  );
   
 };
