+++ src/MainWindow/index.jsx
@@ -4,7 +4,7 @@ import HomePage from "./HomePage";
 import ErrorPage from "./ErrorPage";
 import ScanningPage from "./ScanningPage";
 import ResultPage from "./ResultPage";
-import Button from "../common/components/Button";
+import OnboardingComponent from "./Onboarding/OnboardingComponent";
 import ConnectionNotification from "./ConnectionNotification";
 import services from "../services";
 import "./MainWindow.css";
@@ -12,31 +12,68 @@ import "./MainWindow.css";
 
 const MainWindow = ({ appVersion }) => {
   const [completedScanId, setCompletedScanId] = useState(null);
-  
-  return (
-    <>
-      <ConnectionNotification />
-      <Router>
-        <Routes>
-          <Route
-            path="/"
-            element={
-              <HomePage
-                appVersion={appVersion}
-                setCompletedScanId={setCompletedScanId}
-              />
-            }
-          />
-          <Route path="/scanning" element={<ScanningPage />} />
-          <Route
-            path="/result"
-            element={<ResultPage completedScanId={completedScanId} />}
-          />
-          <Route path="/error" element={<ErrorPage />} />
-        </Routes>
-      </Router>
-    </>
-  );
+  const [email, setEmail] = useState(''); 
+  const [name, setName] = useState('');
+  const [userInputErrorMessage, setUserInputErrorMessage] = useState('');
+  const [dataExistStatus, setDataExistStatus] = useState(null);
+
+  useEffect(() => {
+    window.services.userDataExists((status) => {
+      setDataExistStatus(status)
+    })
+  }, []);
+
+  const handleSetUserData =  (event) =>  {
+    event.preventDefault();
+
+    if (!services.isValidEmail(email)) {
+      setUserInputErrorMessage('Please enter a valid email.'); 
+      return;
+    }
+
+    window.services.setUserData({name: name, email: email, autoSubmit: true});
+    setDataExistStatus("exists");
+  }
+
+  if (dataExistStatus === "doesNotExist") {
+    return (
+      <OnboardingComponent 
+        handleSetUserData={handleSetUserData} 
+        setName={setName}
+        setEmail={setEmail}
+        userInputErrorMessage={userInputErrorMessage}
+      />
+    )
+  }
+
+  if (dataExistStatus === "exists") {
+    return (
+      <>
+        <ConnectionNotification />
+        <Router>
+          <Routes>
+            <Route
+              path="/"
+              element={
+                <HomePage
+                  appVersion={appVersion}
+                  setCompletedScanId={setCompletedScanId}
+                />
+              }
+            />
+            <Route path="/scanning" element={<ScanningPage />} />
+            <Route
+              path="/result"
+              element={<ResultPage completedScanId={completedScanId} />}
+            />
+            <Route path="/error" element={<ErrorPage />} />
+          </Routes>
+        </Router>
+      </>
+    );
+  }
+
+  return null;
 }
